继续初始化知识图谱

+++ ruoyi-modules/ai-bank/src/main/java/org/ruoyi/knowledgegraph/domain/Step.java
@@ -17,7 +17,7 @@ public class Step {
     private String problemId;
 
     @TableField("step_id")
-    private Integer stepId;
+    private Long stepId;
 
     @TableField("operation")
     private String operation;
